[#71084230] Wanted Form
Hi Hes -

I have created the following in:
Wanted Actions x 2 - item controller
Wanted model
3 x wanted views

The good news is the site is not crashing when I run it. The bad news is
the form is not working, even though I tried to link it up in the menu.
I have a note that you need to do something to get this working? There
were a few things in the 'view' files that I need a hand with as I was
not pulling the correct info from the database. Have you set 'Wanted' up
as a field in the database? Lets finish it in class on Wednesday.

Cheers
Marsha